ahgl/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

gondor==1.0.1
psycopg2==2.4.1
PIL==1.1.7
html5lib==0.90
bleach>=1.0
south==0.7.3

django-redis-cache==0.9.2
sorl-thumbnail==11.12
django-social-auth==0.6.1
django-messages==0.4.4

-e git+https://github.com/natmaster/pybbm.git#egg=pybbm